Move boundwiths processing into separate job
